Updated last messages
