+++ innterpret/localization/__init__.py
@@ -2,4 +2,5 @@ from __future__ import absolute_import
 
 # -- IMPORT CLASS METHODS -- #
 from .occlusion import OcclusionMap
-from .gradcam import GradCAM, GuidedGradCAM
+from .gradcam import GradCAM
+from .guidedcam import GuidedGradCAM
