+++ innterpret/localization/occlusion.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 
 # -- IMPORT -- #
-from .. import print_msg
+from .. import __verbose__ as vrb
 from ..utils.data import load_image
 from ..utils.tensor import decode_predictions
 import matplotlib
@@ -11,63 +11,62 @@ import PIL
 import numpy as np
 import math
 
-# -- OCCLUSION MAP METHOD -- #
 class OcclusionMap():
+	""">> CLASS:OCCLUSIONMAP: Method that occludes part of the image and tracks the precision."""
 	def __init__(self,model):
-		print_msg(self.__class__.__name__+' Initializing')
-		print_msg('--------------------------')
+		vrb.print_msg(self.__class__.__name__+' Initializing')
+		vrb.print_msg('--------------------------')
 		self.model = model
 		self.inputSize = model.inputs[0].get_shape()[1]
 		self.numClasses = model.outputs[0].get_shape()[1]-1
-		print_msg('========== DONE ==========\n')
+		vrb.print_msg('========== DONE ==========\n')
 
-	# >> EXECUTE: returns the result of the Occlusion Map method
-	def execute(self,fileName,occSize=15,occStride=3,verbose=True):
-		print_msg(self.__class__.__name__+' Analyzing')
-		print_msg('--------------------------')
-		assert 1 <= occSize <= int(self.inputSize/4)
-		assert 1 <= occStride <= 10
+	def execute(self,fileName,winSize=15,winStride=3):
+		""">> EXECUTE: returns the result of the method."""
+		vrb.print_msg(self.__class__.__name__+' Analyzing')
+		vrb.print_msg('--------------------------')
+		#assert 1 <= winSize <= int(self.inputSize/4)
+		#assert 1 <= winStride <= 10
 		self.rawData = load_image(fileName,preprocess=False)
 		self.imgInput = load_image(fileName)
-		print_msg('Preparing Mask')
-		print_msg('--------------------------')
-		H,W,_ = imgInput[0].shape
-		outH = int(math.ceil((H-occSize)/(occStride+1)))
-		outW = int(math.ceil((W-occSize)/(occStride+1)))
+		vrb.print_msg('Preparing Mask')
+		vrb.print_msg('--------------------------')
+		H,W,_ = self.imgInput[0].shape
+		outH = int(math.ceil((H-winSize)/(winStride+1)))
+		outW = int(math.ceil((W-winSize)/(winStride+1)))
 		heatMap = np.zeros((outH,outW))
-		fullPred = self.model.predict(imgInput)
+		fullPred = self.model.predict(self.imgInput)
 		predictions = decode_predictions(fullPred)
-		print_msg('Top 5 predicted classes: '+str(predictions),option='input')
-		self.maxClass = int(input(print_msg('Select the class to explain (0-'+str(self.numClasses)+'): ',
-			show=False,option='input')))
-		if verbose:
-			print_msg('Prediction of the Selected Class: '+str(fullPred[0][self.maxClass]),option='verbose')
-			print_msg('Number of iterations: '+str(outH*outW),option='verbose')
-			print_msg('Dimensions of Resulting HeatMap: '+str(outH)+'x'+str(outW),option='verbose')
-			k = 1
-			print_msg('------------------',option='verbose')
+		vrb.print_msg('Top 5 predicted classes: '+str(predictions))
+		self.maxClass = int(input(vrb.set_msg('Select the class to explain (0-'+str(self.numClasses)+'): ')))
+		vrb.print_msg('Prediction of the Selected Class: '+str(fullPred[0][self.maxClass]))
+		vrb.print_msg('Number of iterations: '+str(outH*outW))
+		vrb.print_msg('Dimensions of Resulting HeatMap: '+str(outH)+'x'+str(outW))
+		k = 1
+		vrb.print_msg('------------------')
 		for row in range(outH):
 			for col in range(outW):
-				startH = row*occStride; endH = min(H,startH+occSize)
-				startW = col*occStride; endW = min(W,startW+occSize)
-				imgData = imgInput.copy()
+				startH = row*winStride; endH = min(H,startH+winSize)
+				startW = col*winStride; endW = min(W,startW+winSize)
+				imgData = self.imgInput.copy()
 				imgData[:,startH:endH,startW:endW,:] = 0
 				prediction = self.model.predict(imgData)
 				heatMap[row,col] = prediction[0][self.maxClass]
-				if k % 50 == 0 and verbose:
-					print_msg('Result: '+str(self.maxClass)+' ['+str(prediction[0][self.maxClass])+']',option='verbose')
-					print_msg(str(k)+' out of '+str(outH*outW),option='verbose')
-					print_msg('------------------',option='verbose')
+				if k % 50 == 0:
+					vrb.print_msg('Result: '+str(self.maxClass)+' ['+str(prediction[0][self.maxClass])+']')
+					vrb.print_msg(str(k)+' out of '+str(outH*outW))
+					vrb.print_msg('------------------')
 				k += 1
 		self.invHeatMap = 1-heatMap
-		print_msg('========== DONE ==========\n')
+		self.outH = outH; self.outW = outW
+		vrb.print_msg('========== DONE ==========\n')
 		return self.invHeatMap
 
-	# >> VISUALIZE: returns a graph with the results.
 	def visualize(self,savePath,cmap='jet'):
-		print_msg('Visualize '+self.__class__.__name__+' Result...')
-		print_msg('--------------------------')
-		imgRes = self.rawData.resize((outW, outH), PIL.Image.BILINEAR)
+		""">> VISUALIZE: returns a graph with the results."""
+		vrb.print_msg('Visualize '+self.__class__.__name__+' Result...')
+		vrb.print_msg('--------------------------')
+		imgRes = self.rawData.resize((self.outW, self.outH), PIL.Image.BILINEAR)
 		fig = plt.figure(figsize=(6, 4))
 		plt.subplot(121)
 		plt.title('Raw Image')
@@ -79,5 +78,5 @@ class OcclusionMap():
 		plt.imshow(imgRes)
 		plt.imshow(self.invHeatMap,cmap=cmap,interpolation='bilinear',alpha=0.5)
 		fig.savefig(savePath,dpi=250)
-		print_msg('========== DONE ==========\n')
+		vrb.print_msg('========== DONE ==========\n')
