+++ innterpret/signal/__init__.py
@@ -1,5 +1,7 @@
 from __future__ import absolute_import
 
 # -- IMPORT METHOD CLASSES -- #
-from .deconv import Deconvolution, GuidedBackProp
-from .maximization import ActMaximization, FeatMaximization
+from .deconv import Deconvolution
+from .guided import GuidedBackProp
+from .clsmax import ActMaximization
+from .featmax import FeatMaximization
