+++ innterpret/surrogate/__init__.py
@@ -1,4 +1,5 @@
 from __future__ import absolute_import
 
 # -- IMPORT METHOD CLASSES -- #
-from .approximation import GlobalModel, LIMEModel
+from .approx import GlobalModel
+from .lime import LIMEModel
