+++ innterpret/surrogate/approx.py
@@ -1,0 +1,8 @@
+from __future__ import absolute_import
+
+# -- IMPORT -- #
+
+class GlobalModel():
+	""">> CLASS:GLOBALMODEL:"""
+	def __init__(self):
+		raise NotImplementedError
