--- innterpret/surrogate/approximation.py
@@ -1,15 +1,0 @@
-from __future__ import absolute_import
-
-# -- IMPORT -- #
-
-# -- ACTIVATION VISUALIZATION METHOD -- #
-class GlobalModel():
-
-	def __init__(self):
-		raise NotImplementedError
-
-# -- ACTIVATION VISUALIZATION METHOD -- #
-class LIMEModel():
-
-	def __init__(self):
-		raise NotImplementedError
