+++ innterpret/surrogate/lime.py
@@ -1,0 +1,8 @@
+from __future__ import absolute_import
+
+# -- IMPORT -- #
+
+class LIMEModel():
+	""">> CLASS:LIMEMODEL: http://arxiv.org/abs/1602.04938."""
+	def __init__(self):
+		raise NotImplementedError
